requirements.txt
asgiref==3.8.1
gunicorn==21.2.0
packaging==24.0
psycopg2-binary==2.9.9
sqlparse==0.4.4

django==5.0.3
django-cors-headers==3.14.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.2